+++ openmetadata-ui/src/main/resources/ui/src/components/TopicDetails/TopicSchema/TopicSchema.mock.ts
